requirements.txt
requests
PyExecJS
scrapy
PyMySQL
celery===4.4.7
django-mptt
django_redis
selenium
lxml
django-background-tasks